feat(T-unify-02): extract shared configuration types
- Create new config.rs module with StepParams and PipelineCfg
- Move StepParams from world.rs to config.rs
- Move PipelineCfg from pipeline.rs to config.rs
- Update all imports to use config module
- Breaks circular dependency between world.rs and pipeline.rs

All tests pass. Pure refactoring with no behavior change.
