+++ aule/engine/src/pipeline.rs
@@ -7,131 +7,9 @@
 //! - Subduction/transform/rifting/orogeny still edit `depth_m` additively on top of age-derived bathy.
 //!   If these are applied at high cadence without normalization, long-run drift of `depth_m` can occur.
 
-use crate::{boundaries, flexure_loads, ridge, subduction, transforms, world::World};
+use crate::{boundaries, config::PipelineCfg, flexure_loads, ridge, subduction, transforms, world::World};
 use std::thread;
 
-/// Pipeline configuration used by Simple and Advanced modes.
-#[derive(Clone, Copy, Debug)]
-pub struct PipelineCfg {
-    /// Time step in Myr.
-    pub dt_myr: f32,
-    /// Number of steps to run per frame (viewer may loop externally as well).
-    pub steps_per_frame: u32,
-    /// Enable flexure solve stage.
-    pub enable_flexure: bool,
-    /// Enable erosion/diffusion stage.
-    pub enable_erosion: bool,
-    /// Target land fraction (0..1) used by sea-level solve.
-    pub target_land_frac: f32,
-    /// If true, keep eta fixed (skip sea-level regulation this tick).
-    pub freeze_eta: bool,
-    /// If true, append a one-line mass budget log each step.
-    pub log_mass_budget: bool,
-    /// Enable subduction band edits (disable to isolate physics noise)
-    pub enable_subduction: bool,
-    /// Enable rigid plate motion (advect `plate_id`, refresh velocities)
-    pub enable_rigid_motion: bool,
-    /// Cadence controls (>=1); a stage runs when `(step_idx+1) % cadence == 0`.
-    /// Transformer bands cadence in steps.
-    pub cadence_trf_every: u32,
-    /// Subduction bands cadence in steps.
-    pub cadence_sub_every: u32,
-    /// Flexure cadence in steps.
-    pub cadence_flx_every: u32,
-    /// Sea-level (eta) solve cadence in steps.
-    pub cadence_sea_every: u32,
-    /// Surface processes cadence in steps.
-    pub cadence_surf_every: u32,
-    /// Sub-steps for transforms per cadence (narrow operator stability)
-    pub substeps_transforms: u32,
-    /// Sub-steps for subduction band deltas per cadence
-    pub substeps_subduction: u32,
-    /// If true, attempt GPU flexure (experimental). Fallback to Winkler if unavailable.
-    pub use_gpu_flexure: bool,
-    /// GPU flexure: number of multigrid levels.
-    pub gpu_flex_levels: u32,
-    /// GPU flexure: V-cycles per cadence.
-    pub gpu_flex_cycles: u32,
-    /// GPU flexure: weighted-Jacobi omega.
-    pub gpu_wj_omega: f32,
-    /// Subtract mean load before solving (stability, removes rigid-body mode).
-    pub subtract_mean_load: bool,
-    // Subduction tunables (viewer-controlled)
-    /// Convergence threshold in m/yr used for band detection.
-    pub sub_tau_conv_m_per_yr: f32,
-    /// Trench band half-width in km on subducting side.
-    pub sub_trench_half_width_km: f32,
-    /// Arc band peak offset from trench hinge in km (overriding side).
-    pub sub_arc_offset_km: f32,
-    /// Arc band half-width in km.
-    pub sub_arc_half_width_km: f32,
-    /// Back-arc band width in km (behind arc).
-    pub sub_backarc_width_km: f32,
-    /// Trench deepening magnitude in meters (positive deepens).
-    pub sub_trench_deepen_m: f32,
-    /// Arc uplift magnitude in meters (negative uplifts/shallows).
-    pub sub_arc_uplift_m: f32,
-    /// Back-arc uplift magnitude in meters (negative uplifts/shallows).
-    pub sub_backarc_uplift_m: f32,
-    /// Absolute rollback offset applied to overriding distances in meters.
-    pub sub_rollback_offset_m: f32,
-    /// Rollback rate in km/Myr for time-progression (applied by viewer logic).
-    pub sub_rollback_rate_km_per_myr: f32,
-    /// If true, back-arc is deepened (extension mode) instead of uplifted.
-    pub sub_backarc_extension_mode: bool,
-    /// Back-arc deepening magnitude in meters when in extension mode.
-    pub sub_backarc_extension_deepen_m: f32,
-    /// Continental fraction threshold [0,1] for gating trench deepening.
-    pub sub_continent_c_min: f32,
-
-    // Surface processes parameters (viewer-controlled)
-    /// Stream-power coefficient for fluvial incision (yr⁻¹ m^(1-m) s^m units folded)
-    pub surf_k_stream: f32,
-    /// Stream-power m exponent
-    pub surf_m_exp: f32,
-    /// Stream-power n exponent
-    pub surf_n_exp: f32,
-    /// Hillslope diffusion coefficient κ (m²/yr)
-    pub surf_k_diff: f32,
-    /// Sediment transport coefficient (nondimensional scaling)
-    pub surf_k_tr: f32,
-    /// Transport-limited p exponent
-    pub surf_p_exp: f32,
-    /// Transport-limited q exponent
-    pub surf_q_exp: f32,
-    /// Sediment density (kg/m³)
-    pub surf_rho_sed: f32,
-    /// Minimum slope threshold
-    pub surf_min_slope: f32,
-    /// Subcycles per step for numerical stability (>=1)
-    pub surf_subcycles: u32,
-    /// If true, couple surface mass change back into flexure load (placeholder)
-    pub surf_couple_flexure: bool,
-
-    // Plate lifecycle (scaffolding)
-    /// If >0, every N steps attempt to spawn a microplate at a random convergent cell.
-    pub cadence_spawn_plate_every: u32,
-    /// If >0, every N steps attempt to retire the smallest-degree plate into a neighbor.
-    pub cadence_retire_plate_every: u32,
-
-    // Force-balance Euler update
-    /// If >0, every N steps apply a small force-balance update to plate omegas.
-    pub cadence_force_balance_every: u32,
-    /// Force-balance gain (unitless scale to map boundary drive to dω).
-    pub fb_gain: f32,
-    /// Force-balance damping per Myr.
-    pub fb_damp_per_myr: f32,
-    /// Weight for convergent edges.
-    pub fb_k_conv: f32,
-    /// Weight for divergent edges.
-    pub fb_k_div: f32,
-    /// Weight for transform edges.
-    pub fb_k_trans: f32,
-    /// Max per-step Δω clamp.
-    pub fb_max_domega: f32,
-    /// Max absolute |ω| clamp.
-    pub fb_max_omega: f32,
-}
 
 /// Borrowed views of surface fields required by the pipeline.
 pub struct SurfaceFields<'a> {
